Update workflow gem to 1.1.3

+++ dependencies/ruby/build_openstudio_gems.rb
@@ -6,6 +6,11 @@ install_dir = ARGV[0]
 tar_exe = ARGV[1]
 expected_ruby_version = ARGV[2]
 
+def system_call(cmd)
+  puts cmd
+  system(cmd)
+end
+
 ENV['PATH'] = "#{ENV['PATH']}#{File::PATH_SEPARATOR}#{File.dirname(tar_exe)}"
 
 if RUBY_VERSION != expected_ruby_version
@@ -25,13 +30,6 @@ if File.exists?(install_dir)
   FileUtils.rm_rf(install_dir)
 end
 
-ENV['BUNDLE_WITHOUT'] = 'test'
-bundle_exe = File.join(RbConfig::CONFIG['bindir'], 'bundle')
-
-if !File.exists?(bundle_exe)
-  raise "Required bundle executable not found"
-end
-
 bundle_version = nil
 File.open('Gemfile', 'r') do |f|
   while line = f.gets
@@ -43,11 +41,19 @@ end
 if bundle_version.nil?
   raise "Cannot determine bundle version"
 end
-
-system("#{bundle_exe} install --without=test --path='#{install_dir}'")
+bundle_version = bundle_version.gsub(/['=~> ]/, '')
 
 puts "Installing bundler #{bundle_version}"
-system("gem install bundler --version #{bundle_version} --install-dir='#{install_dir}/ruby/#{ruby_gem_dir}'")
+system_call("gem install bundler --version #{bundle_version} --install-dir='#{install_dir}/ruby/#{ruby_gem_dir}'")
+
+ENV['BUNDLE_WITHOUT'] = 'test'
+bundle_exe = File.join("#{install_dir}/ruby/#{ruby_gem_dir}", 'bin', 'bundle')
+
+if !File.exists?(bundle_exe)
+  raise "Required bundle executable not found"
+end
+
+system_call("#{bundle_exe} _#{bundle_version}_ install --without=test --path='#{install_dir}'")
 
 FileUtils.rm_rf("#{install_dir}/ruby/#{ruby_gem_dir}/cache")
 
@@ -93,9 +99,7 @@ FileUtils.cp('Gemfile.lock', "#{install_dir}/.")
 
 Dir.chdir("#{install_dir}/..")
 
-command = "\"#{tar_exe}\" -zcvf \"openstudio-gems-#{DateTime.now.strftime("%Y%m%d")}.tar.gz\" \"openstudio-gems\""
-puts command
-system(command)
+system_call("\"#{tar_exe}\" -zcvf \"openstudio-gems-#{DateTime.now.strftime("%Y%m%d")}.tar.gz\" \"openstudio-gems\"")
 
 # md5sum openstudio-gems-YYYYMMDD.tar.gz
 # upload openstudio-gems-YYYYMMDD.tar.gz to s3, update openstudiocore/CMakeLists.txt
